+++ Brainfuck/src/optimizer.cpp
@@ -111,7 +111,7 @@ namespace bf {
 			//iterator past the last guaranteed instruction
 			auto const last_executable_instruction = opened_loops.empty() ? first_io_operation //if we are not in loop, we end at the IO
 				: [&opened_loops]() { //otherwise find the outermost opened loop and take its opening brace
-				for (int redundant_loops = opened_loops.size() - 1; redundant_loops; --redundant_loops)
+				for (int redundant_loops = static_cast<int>(opened_loops.size()) - 1; redundant_loops; --redundant_loops)
 					opened_loops.pop();
 				return opened_loops.top();
 			}();
@@ -139,12 +139,12 @@ namespace bf {
 			for (cell_t * iter = static_cast<cell_t*>(speculative_emulator.memory_begin()), *end = static_cast<cell_t*>(speculative_emulator.memory_end());;) {
 				cell_t *next = std::find_if(iter, end, [](cell_t cell) {return cell != 0; }); //find the next nonzero cell
 				if (next == end) { //if we run out of cells, move the CPR to the desired position within the memory as if the code had been run normally
-					new_tree.add_instruction(instruction_type::left,
-						std::distance(static_cast<cell_t*>(speculative_emulator.memory_begin()), iter) - speculative_emulator.cell_pointer_offset(), 0);
+					new_tree.add_instruction(instruction_type::left,static_cast<int>(std::distance(static_cast<cell_t*>(speculative_emulator.memory_begin())
+						, iter)) - speculative_emulator.cell_pointer_offset(), 0);
 					break;
 				}
 				//emit instructions to traverse emulator's memory and load constants
-				new_tree.add_instruction(instruction_type::right, std::distance(iter, next) + 1, 0);
+				new_tree.add_instruction(instruction_type::right, static_cast<int>(std::distance(iter, next)) + 1, 0);
 				new_tree.add_instruction(instruction_type::load_const, *next, 0);
 				iter = std::next(next);
 			}
